refactor: 강의 참조 변경
서비스 클래스에서 모델 패키지를 사용하기 위해 강의 가져오기를 리팩토링합니다.
