+++ src/main/java/com/twentyone/steachserver/domain/lecture/repository/LectureRepository.java
@@ -1,6 +1,7 @@
 package com.twentyone.steachserver.domain.lecture.repository;
 
-import com.twentyone.steachserver.domain.lecture.Lecture;
+
+import com.twentyone.steachserver.domain.lecture.model.Lecture;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LectureRepository extends JpaRepository<Lecture, Integer> {
