+++ src/main/java/com/twentyone/steachserver/domain/quiz/service/QuizServiceImpl.java
@@ -1,11 +1,10 @@
 package com.twentyone.steachserver.domain.quiz.service;
 
-import com.twentyone.steachserver.domain.lecture.Lecture;
+import com.twentyone.steachserver.domain.lecture.model.Lecture;
 import com.twentyone.steachserver.domain.lecture.service.LectureService;
 import com.twentyone.steachserver.domain.quiz.dto.QuizRequestDto;
 import com.twentyone.steachserver.domain.quiz.dto.QuizResponseDto;
 import com.twentyone.steachserver.domain.quiz.model.Quiz;
-import com.twentyone.steachserver.domain.quiz.model.QuizChoice;
 import com.twentyone.steachserver.domain.quiz.repository.QuizRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
